% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{any_swearword}
\alias{any_swearword}
\title{Detect if any word is a Swear Word}
\usage{
any_swearword(vec, lang_code)
}
\arguments{
\item{vec}{the vector to test for swear words}

\item{lang_code}{the code of the language to look for}
}
\description{
Detect if any word is a Swear Word
}
\examples{
vec <- c("Yeah nigga, we're still fucking with you, 
Still waters run deep, Still Snoop Dogg and D.R.E")
any_swearword(vec, "en")
}
