% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tools.R
\name{file_lang_code}
\alias{file_lang_code}
\title{Returns language code from file name}
\usage{
file_lang_code(lang_file)
}
\arguments{
\item{lang_file}{Language file name, either absolute
or relative.}
}
\value{
Language code.
}
\description{
Returns language code from file name
}
