% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tools.R
\name{file_lang_name}
\alias{file_lang_name}
\title{Returns language name from file name}
\usage{
file_lang_name(lang_file)
}
\arguments{
\item{lang_file}{Language file name, either absolute
or relative.}
}
\value{
Language name.
}
\description{
Returns language name from file name
}
