% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{get_swearwords}
\alias{get_swearwords}
\title{Get swear words for one language}
\usage{
get_swearwords(lang_code)
}
\arguments{
\item{lang_code}{\href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639-1}
two letter language code.}
}
\value{
Data frame of swear words of one language.
}
\description{
If the language doesn't exist yet, it throws
an error. Otherwise it returns a filtered
\code{swear_words} data frame.
}
\examples{
en_swear_words <- get_swearwords("en")
head(en_swear_words)
}
