% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{is_swearword}
\alias{is_swearword}
\title{Detect if a word is a Swear Word}
\usage{
is_swearword(word, lang_code)
}
\arguments{
\item{word}{the word to look for}

\item{lang_code}{the code of the language to look for}
}
\value{
TRUE if the word is in the database, FALSE if not
}
\description{
Detect if a word is a Swear Word
}
\examples{
is_swearword("shit", "en")
}
