% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tools.R
\name{load_lang_from_file}
\alias{load_lang_from_file}
\title{Loads a single language data frame from file}
\usage{
load_lang_from_file(lang_file)
}
\arguments{
\item{lang_file}{Language file name with full path.}
}
\value{
Data frame of swear words in one language.
}
\description{
Loads a single language data frame from file
}
