% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tools.R
\name{load_langs}
\alias{load_langs}
\title{Create a summary df with languages and their counts}
\usage{
load_langs()
}
\value{
Data frame with language codes, language names,
  word counts and a formatted markdown table row.
}
\description{
Create a summary df with languages and their counts
}
