% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{n_swearword}
\alias{n_swearword}
\title{How muck are Swear Words?}
\usage{
n_swearword(vec, lang_code)
}
\arguments{
\item{vec}{the vector to test for swear words}

\item{lang_code}{the code of the language to look for}
}
\value{
the number of swear words detected
}
\description{
How muck are Swear Words?
}
\examples{
vec <- c("Yeah nigga, we're still fucking with you, 
Still waters run deep, Still Snoop Dogg and D.R.E")
n_swearword(vec, "en")
}
