% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{pct_swearword}
\alias{pct_swearword}
\title{How much are Swear Words?}
\usage{
pct_swearword(vec, lang_code)
}
\arguments{
\item{vec}{the vector to test for swear words}

\item{lang_code}{the code of the language to look for}
}
\value{
the percentage of swear words detected
}
\description{
How much are Swear Words?
}
\examples{
vec <- c("Yeah nigga, we're still fucking with you, 
Still waters run deep, Still Snoop Dogg and D.R.E")
pct_swearword(vec, "en")
}
