% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tools.R
\name{split_lang_file}
\alias{split_lang_file}
\title{Splits lang file name in language code and name}
\usage{
split_lang_file(lang_file)
}
\arguments{
\item{lang_file}{Language file name, either absolute
or relative.}
}
\value{
Character vector of length 2. First
  element is language code, second element
  is language name.
}
\description{
Splits lang file name in language code and name
}
