% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweary.R
\docType{data}
\name{swear_words}
\alias{swear_words}
\title{Data frame of swear words}
\format{A data frame with 96 rows and 2 variables.}
\usage{
swear_words
}
\description{
A dataset containing a list of swear words across multiple
  languages.
}
\details{
\describe{
  \item{word}{A swear word.}
  \item{language}{A character code of the language the swear word
    is recorded in.}
}
}
\keyword{datasets}
