% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{sweary_censor}
\alias{sweary_censor}
\title{Censor a character vector}
\usage{
sweary_censor(vec, lang_code, with = "***")
}
\arguments{
\item{vec}{a vector to censor}

\item{lang_code}{the code of the language to look for}

\item{with}{the element to use as a censor mark}
}
\value{
a censored version of the vector
}
\description{
Censor a character vector
}
\examples{
vec <- c("Yeah, nigga, we're still fucking with you, Still waters run deep ,Still Snoop Dogg and D.R.E. '99, nigga, guess who's back; Still, still doin' that shit, Andre? Oh for sho' Yeah! Check me out")
sweary_censor(vec, "en")
}
