% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{sweary_pct}
\alias{sweary_pct}
\title{How much are Swear Words?}
\usage{
sweary_pct(vec, lang_code)
}
\arguments{
\item{vec}{the vector to test for swear words}

\item{lang_code}{the code of the language to look for}
}
\value{
the percentage of swear words detected
}
\description{
How much are Swear Words?
}
\examples{
vec <- c("Yeah, nigga, we're still fucking with you, Still waters run deep ,Still Snoop Dogg and D.R.E. '99, nigga, guess who's back; Still, still doin' that shit, Andre? Oh for sho' Yeah! Check me out")
sweary_pct(vec, "en")
}
