% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweary.R
\docType{package}
\name{sweary-package}
\alias{sweary}
\alias{sweary-package}
\title{sweary: A database of swear words from multiple languages}
\description{
sweary: A database of swear words from multiple languages
}
\author{
\strong{Maintainer}: Patrik Drhlik \email{patrik.drhlik@gmail.com}

}
