% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweary.R
\docType{package}
\name{sweary-package}
\alias{sweary}
\alias{sweary-package}
\title{sweary: A database of swear words from multiple languages}
\description{
The packages tries to offer an extensive list of swear
    words from different languages, cherry picked by native speakers.
    It should then be used in different text analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pdrhlik/sweary}
  \item Report bugs at \url{https://github.com/pdrhlik/sweary/issues}
}

}
\author{
\strong{Maintainer}: Patrik Drhlik \email{patrik.drhlik@gmail.com}

}
